Improve styling for NS local unit form
