+++ .changeset/olive-dolphins-wonder.md
@@ -1,0 +1,5 @@
+---
+"go-web-app": patch
+---
+
+Add local unit form
