--- app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/AddEditLocalUnitsModal/AddEditLocalUnitsForm/i18n.json
@@ -1,64 +1,0 @@
-{
-    "namespace": "nationalSocietyLocalUnits",
-    "strings": {
-        "localUnitsFormSubtype": "Sub-type",
-        "localUnitsFormSubtypeDescription": "(e.g. hospital)",
-        "localUnitsFormEnglishDescription": "(English)",
-        "localUnitsFormLocalDescription": "(Local)",
-        "localUnitsFormLocalUnitName": "Local unit name",
-        "localUnitsFormCoverage": "Coverage",
-        "localUnitsFormFocalPerson": "Focal person",
-        "localUnitsFormSource": "Source",
-        "localUnitsAddressTitle": "Address and Contact",
-        "localUnitsFacilityCategoryTitle": "Facility Category",
-        "localUnitsFacilityCapacityTitle": "Facility Capacity",
-        "localUnitsHumanResourcesTitle": "Human Resources",
-        "localUnitsServicesTitle": "Services",
-        "localUnitsFormAddress": "Address",
-        "localUnitsFormLocality": "Locality",
-        "localUnitsFormPostCode": "Post code",
-        "localUnitsFormPhone": "Local unit phone number",
-        "localUnitsFormEmail": "Local unit email",
-        "localUnitsFormWebsite": "Local unit website",
-        "localUnitsFormType": "Local unit type",
-        "localUnitsVisibility": "Visibility",
-        "localUnitsFormAffiliation": "Affiliation",
-        "localUnitsFormOtherAffiliation": "Other Affiliation",
-        "localUnitsFormFunctionality": "Functionality",
-        "localUnitsFormHealthFacilityType": "Health Facility type",
-        "localUnitsFormOtherFacilityType": "Other Facility type",
-        "localUnitsFormHospitalType": "Hospital type",
-        "localUnitsFormTeachingHospital": "Teaching Hospital",
-        "localUnitsFormInPatientCapacity": "In-patient capacity",
-        "localUnitsFormIsolationRoomsWards": "Isolation rooms wards",
-        "localUnitsFormFocalPointName": "Focal person name",
-        "localUnitsFormFocalPointPosition": "Focal person position",
-        "localUnitsFormFocalPointEmail": "Focal person email",
-        "localUnitsFormFocalPointPhoneNumber": "Focal person phone number",
-        "localUnitsFormSpecialties": "Specialties",
-        "localUnitsFormOtherServices": "Other Services",
-        "localUnitsFormBloodServices": "Blood Services",
-        "localUnitsFormProfessionalTrainingFacilities": "Professional training facilities",
-        "localUnitsFormGeneralMedicalServices": "General medical services",
-        "localUnitsFormSpecialist": "Specialist",
-        "localUnitsFormSpecializedMedicalService": "Specialized medical beyond primary level",
-        "localUnitsFormMaximumCapacity": "Maximum capacity",
-        "localUnitsFormNumberOfIsolationRooms": "Number of isolation rooms",
-        "localUnitsFormWarehousing": "Warehousing",
-        "localUnitsFormColdChain": "Cold chain",
-        "localUnitsFormAmbulanceTypeA": "Ambulance Type A",
-        "localUnitsFormAmbulanceTypeB": "Ambulance Type B",
-        "localUnitsFormAmbulanceTypeC": "Ambulance Type C",
-        "localUnitsFormTotalNumberOfHumanResources": "Total number of human resource",
-        "localUnitsFormGeneralPractitioner": "General Practitioner",
-        "localUnitsFormResidentsDoctor": "Residents doctor",
-        "localUnitsFormNurse": "Nurse",
-        "localUnitsFormDentist": "Dentist",
-        "localUnitsFormNursingAid": "Nursing aid",
-        "localUnitsFormMidwife": "Midwife",
-        "localUnitsFormOtherMedicalHeal": "Other medical heal",
-        "localUnitsFormOtherProfiles": "Other profiles",
-        "localUnitsFormCommentsNS": "Comments by the NS",
-        "localUnitsFormSubmitButtonLabel": "Submit"
-    }
-}
