--- app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/AddEditLocalUnitsModal/AddEditLocalUnitsForm/styles.module.css
@@ -1,7 +1,0 @@
-.local-units-form {
-    width: 100%;
-
-    .point-input-map {
-        height: 40rem;
-    }
-}
