--- app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/AddEditLocalUnitsModal/i18n.json
@@ -1,7 +1,0 @@
-{
-    "namespace": "nationalSocietyLocalUnits",
-    "strings": {
-        "localUnitsModalHeading": "Local Unit - {type}",
-        "localUnitsModalContent": "This is the content"
-    }
-}
