--- app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/AddEditLocalUnitsModal/styles.module.css
@@ -1,3 +1,0 @@
-.add-edit-modal {
-    display: flex;
-}
