+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/LocalUnitsFormModal/LocalUnitsForm/i18n.json
@@ -1,0 +1,65 @@
+{
+    "namespace": "",
+    "strings": {
+        "subtype": "Sub-type",
+        "subtypeDescription": "(e.g. hospital)",
+        "englishLabel": "English",
+        "localLabel": "Local",
+        "localUnitName": "Local unit name",
+        "coverage": "Coverage",
+        "focalPerson": "Focal person name",
+        "focalPointLabel": "Focal point",
+        "source": "Source",
+        "addressAndContactTitle": "Address and Contact",
+        "facilityCategoryTitle": "Facility Category",
+        "facilityCapacityTitle": "Facility Capacity",
+        "humanResourcesTitle": "Human Resources",
+        "servicesTitle": "Services",
+        "address": "Address",
+        "locality": "Locality",
+        "postCode": "Local unit post code",
+        "phone": "Local unit phone number",
+        "email": "Local unit email",
+        "website": "Local unit website",
+        "type": "Local unit type",
+        "visibility": "Visibility",
+        "affiliation": "Affiliation",
+        "otherAffiliation": "Other Affiliation",
+        "functionality": "Functionality",
+        "healthFacilityType": "Health Facility type",
+        "otherFacilityType": "Other Facility type",
+        "hospitalType": "Hospital type",
+        "teachingHospital": "Teaching Hospital",
+        "inPatientCapacity": "In-patient capacity",
+        "isolationRoomsWards": "Isolation rooms wards",
+        "focalPointName": "Name",
+        "focalPointPosition": "Position",
+        "focalPointEmail": "Email",
+        "focalPointPhoneNumber": "Phone number",
+        "specialties": "Specialties",
+        "otherServices": "Other Services",
+        "bloodServices": "Blood Services",
+        "professionalTrainingFacilities": "Professional training facilities",
+        "generalMedicalServices": "General medical services",
+        "specialist": "Specialist",
+        "specializedMedicalService": "Specialized medical beyond primary level",
+        "maximumCapacity": "Maximum capacity",
+        "numberOfIsolationRooms": "Number of isolation rooms",
+        "warehousing": "Warehousing",
+        "coldChain": "Cold chain",
+        "ambulanceTypeA": "Ambulance Type A",
+        "ambulanceTypeB": "Ambulance Type B",
+        "ambulanceTypeC": "Ambulance Type C",
+        "totalNumberOfHumanResources": "Total number of human resource",
+        "generalPractitioner": "General Practitioner",
+        "residentsDoctor": "Residents doctor",
+        "nurse": "Nurse",
+        "dentist": "Dentist",
+        "nursingAid": "Nursing aid",
+        "midwife": "Midwife",
+        "otherMedicalHeal": "Other medical heal",
+        "otherProfiles": "Other profiles",
+        "commentsNS": "Comments by the NS",
+        "submitButtonLabel": "Submit"
+    }
+}
