+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/LocalUnitsFormModal/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "nationalSocietyLocalUnits",
+    "strings": {
+        "localUnitsModalHeading": "Local Unit Form"
+    }
+}
