+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/LocalUnitsFormModal/styles.module.css
@@ -1,0 +1,3 @@
+.local-units-form-modal {
+    display: flex;
+}
