+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/styles.module.css
@@ -1,4 +1,4 @@
-.national-society-local-units-map {
+.national-society-local-units {
     .content {
         position: relative;
     }
